WebClassLoaderRegistry is introduced: collection of  WebClassLoaders associated with URLs.
As a result, a resource inside of one classloader can be referenced externally (from WebFX browser) and correct classloader will be resolved for it (before this only root resource could be referenced).

wfx:// and java:// protocols implementation was moved to the client (they do not depend on WebFX project now).